stream detail page wip
stream detail view wip
