+++ packages/www/components/Dashboard/Spinner/index.tsx
@@ -1,0 +1,26 @@
+import { keyframes, Box } from "@livepeer.com/design-system";
+
+const rotate = keyframes({
+  "100%": { transform: "rotate(360deg)" },
+});
+
+const Index = ({ css = {}, speed = "1s" }) => (
+  <Box
+    css={{
+      color: "$violet11",
+      border: "2px solid",
+      borderColor: "$loContrast",
+      borderRadius: "50%",
+      borderTopColor: "inherit",
+      width: 26,
+      height: 26,
+      maxWidth: 26,
+      maxHeight: 26,
+      animation: `${rotate} ${speed} linear`,
+      animationIterationCount: "infinite",
+      ...css,
+    }}
+  />
+);
+
+export default Index;
