+++ packages/www/pages/dashboard/streams/[id].tsx
@@ -1,0 +1,1003 @@
+import Link from "next/link";
+import ReactTooltip from "react-tooltip";
+import {
+  Box,
+  Button,
+  Flex,
+  Heading,
+  Grid,
+  Link as A,
+  Radio,
+  Alert,
+} from "@livepeer.com/design-system";
+import Layout from "../../../layouts/dashboard";
+import useLoggedIn from "../../../hooks/use-logged-in";
+import { Stream } from "@livepeer.com/api";
+import { CopyToClipboard } from "react-copy-to-clipboard";
+import { useRouter } from "next/router";
+import Router from "next/router";
+import { useApi, usePageVisibility } from "../../../hooks";
+import { useEffect, useState } from "react";
+import StreamSessionsTable from "components/Dashboard/SessionsTable";
+import DeleteStreamModal from "components/DeleteStreamModal";
+import ConfirmationModal from "components/ConfirmationModal";
+import Modal from "components/Modal";
+import {
+  pathJoin,
+  isStaging,
+  isDevelopment,
+  formatNumber,
+} from "../../../lib/utils";
+import { RenditionsDetails } from "components/StreamsTable";
+import { RelativeTime } from "components/CommonAdminTable";
+import {
+  Cross1Icon as Close,
+  CopyIcon as Copy,
+  QuestionMarkCircledIcon as Help,
+} from "@radix-ui/react-icons";
+import Spinner from "components/Dashboard/Spinner";
+import Player from "components/Dashboard/Player";
+
+type TimedAlertProps = {
+  text: string;
+  close: Function;
+  variant?: string;
+};
+
+const TimedAlert = ({ text, variant, close }: TimedAlertProps) => {
+  const isShown = !!text;
+  const closeVariant = variant ? `close-${variant}` : "close";
+  useEffect(() => {
+    if (isShown) {
+      const interval = setTimeout(() => {
+        close();
+      }, 3000);
+      return () => clearTimeout(interval);
+    }
+  }, [text]);
+  return isShown ? (
+    <Alert css={{ m: 2 }}>{text}</Alert>
+  ) : // <Alert variant={variant} css={{ m: 2 }}>
+  //   {text}
+  //   <Close variant={closeVariant} ml="auto" mr={-2} onClick={() => close()} />
+  // </Alert>
+  null;
+};
+
+type ShowURLProps = {
+  text: string;
+  url: string;
+  anchor?: boolean;
+  urlToCopy?: string;
+};
+
+const ShowURL = ({ text, url, urlToCopy, anchor = false }: ShowURLProps) => {
+  const [isCopied, setCopied] = useState(0);
+  useEffect(() => {
+    if (isCopied) {
+      const interval = setTimeout(() => {
+        setCopied(0);
+      }, isCopied);
+      return () => clearTimeout(interval);
+    }
+  }, [isCopied]);
+  const ccurl = urlToCopy ? urlToCopy : url;
+  return (
+    <Flex css={{ justifyContent: "flex-start", alignItems: "center" }}>
+      {text ? (
+        <Box css={{ minWidth: 125, fontSize: 12, paddingRight: "1em" }}>
+          {text}:
+        </Box>
+      ) : null}
+      <CopyToClipboard text={ccurl} onCopy={() => setCopied(2000)}>
+        <Flex css={{ alignItems: "center" }}>
+          {anchor ? (
+            <a
+              css={{ fontSize: 12, fontFamily: "monospace", mr: 1 }}
+              href={url}
+              target="_blank">
+              {url}
+            </a>
+          ) : (
+            <span css={{ fontSize: 12, fontFamily: "monospace", mr: 1 }}>
+              {url}
+            </span>
+          )}
+          <Copy
+            css={{
+              mr: 1,
+              cursor: "pointer",
+              width: 14,
+              height: 14,
+              color: "$hiContrast",
+            }}
+          />
+        </Flex>
+      </CopyToClipboard>
+      {!!isCopied && (
+        <Box css={{ fontSize: 12, color: "$hiContrast" }}>Copied</Box>
+      )}
+    </Flex>
+  );
+};
+
+const Cell = ({ children }) => {
+  return <Box css={{ mb: "$3" }}>{children}</Box>;
+};
+
+const ClipBut = ({ text }) => {
+  const [isCopied, setCopied] = useState(0);
+  useEffect(() => {
+    if (isCopied) {
+      const timeout = setTimeout(() => {
+        setCopied(0);
+      }, isCopied);
+      return () => clearTimeout(timeout);
+    }
+  }, [isCopied]);
+  return (
+    <CopyToClipboard text={text} onCopy={() => setCopied(2000)}>
+      <Flex
+        css={{
+          alignItems: "center",
+          cursor: "pointer",
+          ml: 0,
+          mr: 0,
+        }}>
+        <Box css={{ mr: 2 }}>{text}</Box>
+        <Copy
+          css={{
+            mr: 1,
+            width: 14,
+            height: 14,
+            color: "$hiContrast",
+          }}
+        />
+        {!!isCopied && (
+          <Box css={{ fontSize: 12, color: "$hiContrast" }}>Copied</Box>
+        )}
+      </Flex>
+    </CopyToClipboard>
+  );
+};
+
+const ID = () => {
+  useLoggedIn();
+  const {
+    user,
+    logout,
+    getStream,
+    deleteStream,
+    getIngest,
+    setRecord,
+    getAdminStreams,
+    terminateStream,
+    suspendStream,
+  } = useApi();
+  const userIsAdmin = user && user.admin;
+  const router = useRouter();
+  const { query } = router;
+  const id = query.id;
+  const [stream, setStream] = useState(null);
+  const [ingest, setIngest] = useState([]);
+  const [deleteModal, setDeleteModal] = useState(false);
+  const [terminateModal, setTerminateModal] = useState(false);
+  const [suspendModal, setSuspendModal] = useState(false);
+  const [notFound, setNotFound] = useState(false);
+  const [recordOffModal, setRecordOffModal] = useState(false);
+  const [isCopied, setCopied] = useState(0);
+  const [lastSession, setLastSession] = useState(null);
+  const [lastSessionLoading, setLastSessionLoading] = useState(false);
+  const [regionalUrlsVisible, setRegionalUrlsVisible] = useState(false);
+  const [resultText, setResultText] = useState("");
+  const [alertText, setAlertText] = useState("");
+  const [videoExists, setVideoExists] = useState<boolean>(false);
+
+  useEffect(() => {
+    if (user && user.admin && stream && !lastSessionLoading) {
+      setLastSessionLoading(true);
+      getAdminStreams({
+        sessionsonly: true,
+        limit: 1,
+        order: "createdAt-true",
+        filters: [{ id: "parentId", value: stream.id }],
+      })
+        .then((res) => {
+          const [streamsOrError] = res;
+          if (Array.isArray(streamsOrError) && streamsOrError.length > 0) {
+            setLastSession(streamsOrError[0]);
+          }
+        })
+        .catch((e) => console.log(e))
+        .finally(() => setLastSessionLoading(false));
+    }
+  }, [user, stream]);
+
+  useEffect(() => {
+    if (isCopied) {
+      const timeout = setTimeout(() => {
+        setCopied(0);
+      }, isCopied);
+      return () => clearTimeout(timeout);
+    }
+  }, [isCopied]);
+
+  useEffect(() => {
+    getIngest(true)
+      .then((ingest) => {
+        if (Array.isArray(ingest)) {
+          ingest.sort((a, b) => a.base.localeCompare(b.base));
+        }
+        setIngest(ingest);
+      })
+      .catch((err) => console.error(err)); // todo: surface this
+  }, [id]);
+  useEffect(() => {
+    if (!id) {
+      return;
+    }
+    getStream(id)
+      .then((stream) => setStream(stream))
+      .catch((err) => {
+        if (err && err.status === 404) {
+          setNotFound(true);
+        }
+        console.error(err);
+      }); // todo: surface this
+  }, [id]);
+  const isVisible = usePageVisibility();
+  useEffect(() => {
+    if (!isVisible || !id || notFound) {
+      return;
+    }
+    const interval = setInterval(() => {
+      getStream(id)
+        .then((stream) => setStream(stream))
+        .catch((err) => console.error(err)); // todo: surface this
+    }, 5000);
+    return () => clearInterval(interval);
+  }, [id, isVisible]);
+  const [keyRevealed, setKeyRevealed] = useState(false);
+  const close = () => {
+    setSuspendModal(false);
+    setTerminateModal(false);
+    setDeleteModal(false);
+    setRecordOffModal(false);
+  };
+
+  if (!user || user.emailValid === false) {
+    return <Layout />;
+  }
+
+  const getIngestURL = (
+    stream: Stream,
+    showKey: boolean,
+    i: number
+  ): string => {
+    const key = showKey ? stream.streamKey : "";
+    return i < ingest.length ? pathJoin(ingest[i].ingest, key) : key || "";
+  };
+  const getPlaybackURL = (stream: Stream, i: number): string => {
+    return i < ingest.length
+      ? pathJoin(ingest[i].base, "hls", `${stream.playbackId}/index.m3u8`)
+      : stream.playbackId || "";
+  };
+  const doSetRecord = async (stream: Stream, record: boolean) => {
+    console.log(`do set record ${stream.id} record ${record}`);
+    setStream(null); // shows 'loading wheel' immediately
+    await setRecord(stream.id, record);
+    setStream(null); // make sure that we will load updated stream
+  };
+
+  const isAdmin = query.admin === "true";
+  const _stream = stream || {};
+  let { broadcasterHost, region } = _stream;
+  if (!broadcasterHost && lastSession && lastSession.broadcasterHost) {
+    broadcasterHost = lastSession.broadcasterHost;
+  }
+  if (!region && lastSession && lastSession.region) {
+    region = lastSession.region;
+  }
+  let broadcasterPlaybackUrl;
+  const playbackId = (stream || {}).playbackId || "";
+  const domain = isStaging() ? "monster" : "com";
+  const globalIngestUrl = `rtmp://rtmp.livepeer.${domain}/live`;
+  const globalPlaybackUrl = `https://cdn.livepeer.${domain}/hls/${playbackId}/index.m3u8`;
+
+  if (stream && stream.region && !lastSession) {
+    broadcasterPlaybackUrl = `https://${stream.region}.livepeer.${domain}/stream/${stream.id}.m3u8`;
+  } else if (lastSession && lastSession.region) {
+    broadcasterPlaybackUrl = `https://${lastSession.region}.livepeer.${domain}/stream/${playbackId}.m3u8`;
+  }
+
+  return (
+    <Layout
+      id="streams"
+      breadcrumbs={[
+        { title: "Streams", href: "/dashboard/streams" },
+        { title: stream?.name },
+      ]}>
+      <Box css={{ p: "$6" }}>
+        {suspendModal && stream && (
+          <ConfirmationModal
+            actionText="Confirm"
+            onClose={close}
+            onAction={() => {
+              const newValue = !stream.suspended;
+              suspendStream(stream.id, !stream.suspended)
+                .then((res) => {
+                  stream.suspended = newValue;
+                  setStream({ ...stream, suspended: newValue });
+                })
+                .catch((e) => {
+                  console.error(e);
+                })
+                .finally(close);
+            }}>
+            {!stream.suspended
+              ? `Are you sure you want to suspend and block this stream? 
+            Any active stream sessions will immediately end. 
+            New sessions will be prevented from starting until unchecked.`
+              : `Are you sure you want to allow new stream sessions again?`}
+          </ConfirmationModal>
+        )}
+        {terminateModal && stream && (
+          <ConfirmationModal
+            actionText="Terminate"
+            onClose={close}
+            onAction={() => {
+              terminateStream(stream.id)
+                .then((res) => {
+                  setResultText(`sucess: ${res}`);
+                })
+                .catch((e) => {
+                  console.error(e);
+                  setAlertText(`${e}`);
+                })
+                .finally(close);
+            }}>
+            Are you sure you want to terminate (stop running live) stream{" "}
+            <b>{stream.name}</b>? Terminating a stream will break RTMP
+            connection.
+          </ConfirmationModal>
+        )}
+        {deleteModal && stream && (
+          <DeleteStreamModal
+            streamName={stream.name}
+            onClose={close}
+            onDelete={() => {
+              deleteStream(stream.id).then(() => Router.replace("/app/user"));
+            }}
+          />
+        )}
+        {recordOffModal && stream && (
+          <Modal onClose={close}>
+            <h3>Are you sure you want to turn off recoding?</h3>
+            <p>
+              Future stream sessions will not be recorded. In progress stream
+              sessions will be recorded. Past sessions recordings will still be
+              available.
+            </p>
+            <Flex css={{ justifyContent: "flex-end" }}>
+              <Button
+                type="button"
+                variant="violet"
+                onClick={close}
+                css={{ mr: 2 }}>
+                Cancel
+              </Button>
+              <Button
+                type="button"
+                variant="violet"
+                onClick={() => {
+                  close();
+                  doSetRecord(stream, false);
+                }}>
+                Turn off recording
+              </Button>
+            </Flex>
+          </Modal>
+        )}
+
+        {stream ? (
+          <>
+            <Flex
+              justify="between"
+              align="end"
+              css={{
+                borderBottom: "1px solid",
+                borderColor: "$slate6",
+                pb: "$3",
+                mb: "$5",
+                width: "100%",
+              }}>
+              <Heading size="2">
+                <Flex>
+                  <Box
+                    css={{
+                      fontWeight: 600,
+                      letterSpacing: "0",
+                    }}>
+                    {stream.name}
+                  </Box>
+                </Flex>
+              </Heading>
+            </Flex>
+            <Grid gap="8" columns="2">
+              <Box>
+                <Heading size="1" css={{ mb: "$3" }}>
+                  Details
+                </Heading>
+                <Flex
+                  css={{
+                    justifyContent: "flex-start",
+                    alignItems: "baseline",
+                    flexDirection: "column",
+                  }}>
+                  <Box
+                    css={{
+                      display: "grid",
+                      alignItems: "center",
+                      gridTemplateColumns: "10em auto",
+                      width: "100%",
+                      fontSize: "$2",
+                      position: "relative",
+                    }}>
+                    <Cell>Stream name</Cell>
+                    <Cell>{stream.name}</Cell>
+                    <Cell>Stream ID</Cell>
+                    <Cell>
+                      <ClipBut text={stream.id}></ClipBut>
+                    </Cell>
+                    <Cell>Stream key</Cell>
+                    <Cell>
+                      {keyRevealed ? (
+                        <Flex>
+                          {stream.streamKey}
+                          <CopyToClipboard
+                            text={stream.streamKey}
+                            onCopy={() => setCopied(2000)}>
+                            <Flex
+                              css={{
+                                alignItems: "center",
+                                cursor: "pointer",
+                                ml: 1,
+                              }}>
+                              <Copy
+                                css={{
+                                  mr: 1,
+                                  width: 14,
+                                  height: 14,
+                                  color: "$hiContrast",
+                                }}
+                              />
+                              {!!isCopied && (
+                                <Box
+                                  css={{ fontSize: 12, color: "$hiContrast" }}>
+                                  Copied
+                                </Box>
+                              )}
+                            </Flex>
+                          </CopyToClipboard>
+                        </Flex>
+                      ) : (
+                        <Button
+                          type="button"
+                          variant="violet"
+                          onClick={() => setKeyRevealed(true)}
+                          css={{ mr: 0, py: "4px" }}>
+                          Reveal secret stream key
+                        </Button>
+                      )}
+                    </Cell>
+                    <Cell>RTMP ingest URL</Cell>
+                    <Cell>
+                      <ShowURL text="" url={globalIngestUrl} anchor={true} />
+                    </Cell>
+                    <Cell>Playback URL</Cell>
+                    <Cell>
+                      <ShowURL text="" url={globalPlaybackUrl} anchor={true} />
+                    </Cell>
+                    <Box
+                      css={{
+                        mt: "$1",
+                        mb: "0",
+                        gridColumn: "1/-1",
+                      }}>
+                      <Box
+                        onClick={() =>
+                          setRegionalUrlsVisible(!regionalUrlsVisible)
+                        }
+                        css={{
+                          cursor: "pointer",
+                          display: "inline-block",
+                          transform: regionalUrlsVisible
+                            ? "rotate(90deg)"
+                            : "rotate(0deg)",
+                          transition: "transform 0.4s ease",
+                        }}>
+                        ▶
+                      </Box>{" "}
+                      Regional ingest and playback URL pairs
+                    </Box>
+                    <Box
+                      css={{
+                        gridColumn: "1/-1",
+                        position: "relative",
+                        overflow: "hidden",
+                        mb: "0.8em",
+                      }}>
+                      <Box
+                        css={{
+                          position: "relative",
+                          overflow: "hidden",
+                          transition: "margin-bottom .4s ease",
+                          mb: regionalUrlsVisible ? "0" : "-100%",
+                          display: "grid",
+                          alignItems: "center",
+                          gridTemplateColumns: "10em auto",
+                        }}>
+                        <Box
+                          css={{
+                            mx: "$2",
+                            mt: "$3",
+                            gridColumn: "1/-1",
+                          }}>
+                          The global RTMP ingest and playback URL pair above
+                          auto detects livestreamer and viewer locations to
+                          provide the optimal Livepeer.com experience.
+                          <Link
+                            href="/docs/guides/dashboard/ingest-playback-url-pair"
+                            passHref>
+                            <A target="_blank">
+                              <i>
+                                Learn more about forgoing the global ingest and
+                                playback URLs before selecting a regional URL
+                                pair.
+                              </i>
+                            </A>
+                          </Link>
+                        </Box>
+                        {/* {!ingest.length && (
+                      <Spinner css={{ mb: 3, width: 32, height: 32 }} />
+                    )} */}
+                        {ingest.map((_, i) => {
+                          return (
+                            <>
+                              <Cell>RTMP ingest URL {i + 1}</Cell>
+                              <Cell>
+                                <ShowURL
+                                  text=""
+                                  url={getIngestURL(stream, false, i)}
+                                  urlToCopy={getIngestURL(stream, false, i)}
+                                  anchor={false}
+                                />
+                              </Cell>
+                              <Box
+                                css={{
+                                  m: "0.4em",
+                                  mb: "1.4em",
+                                }}>
+                                Playback URL {i + 1}
+                              </Box>
+                              <Box
+                                css={{
+                                  m: "0.4em",
+                                  mb: "1.4em",
+                                }}>
+                                <ShowURL
+                                  text=""
+                                  url={getPlaybackURL(stream, i)}
+                                  anchor={true}
+                                />
+                              </Box>
+                            </>
+                          );
+                        })}
+                      </Box>
+                    </Box>
+                    <Box css={{ m: "0.4em", gridColumn: "1/-1" }}>
+                      <hr />
+                    </Box>
+                    <Cell>Record sessions</Cell>
+                    <Box
+                      css={{
+                        m: "0.4em",
+                        justifySelf: "flex-start",
+                        cursor: "pointer",
+                      }}>
+                      <Flex
+                        css={{
+                          alignItems: "flex-start",
+                          justifyItems: "center",
+                        }}>
+                        {/* <Label
+                      onClick={() => {
+                        if (!stream.record) {
+                          doSetRecord(stream, true);
+                        }
+                      }}>
+                      <Radio
+                        autocomplete="off"
+                        name="record-mode"
+                        value={`${!!stream.record}`}
+                        checked={!!stream.record}
+                      />
+                      <Flex css={{ alignItems: "center" }}>On</Flex>
+                    </Label>
+                    <Label css={{ ml: "0.5em" }}>
+                      <Radio
+                        autocomplete="off"
+                        name="record-mode"
+                        value={`${!stream.record}`}
+                        checked={!stream.record}
+                        onClick={(e) => {
+                          if (stream.record) {
+                            setRecordOffModal(true);
+                          }
+                        }}
+                      />
+                      <Flex css={{ alignItems: "center" }}>Off</Flex>
+                    </Label> */}
+                        <Flex
+                          css={{
+                            ml: "0.5em",
+                            minWidth: "24px",
+                            height: "24px",
+                            alignItems: "center",
+                          }}>
+                          <Help
+                            data-tip
+                            data-for={`tooltip-record-${stream.id}`}
+                            css={{
+                              color: "muted",
+                              cursor: "pointer",
+                              ml: 1,
+                              width: "18px",
+                              height: "18px",
+                            }}
+                          />
+                        </Flex>
+                      </Flex>
+                      <ReactTooltip
+                        id={`tooltip-record-${stream.id}`}
+                        className="tooltip"
+                        place="top"
+                        type="dark"
+                        effect="solid">
+                        <p>
+                          When checked, transcoded streaming sessions will be
+                          recorded and stored by Livepeer.com.
+                          <br /> Each recorded session will have a recording
+                          .m3u8 URL for playback. <br />
+                          This feature is currently in beta and free.
+                        </p>
+                      </ReactTooltip>
+                    </Box>
+                    <Box css={{ m: "0.4em", gridColumn: "1/-1" }}>
+                      <hr />
+                    </Box>
+                    <Cell>Suspend and block</Cell>
+                    <Box
+                      css={{
+                        m: "0.4em",
+                        justifySelf: "flex-start",
+                        cursor: "pointer",
+                      }}>
+                      <Flex
+                        css={{
+                          alignItems: "flex-start",
+                          justifyItems: "center",
+                        }}>
+                        {/* <Label
+                      onClick={() => {
+                        if (!stream.suspended) {
+                          setSuspendModal(true);
+                        }
+                      }}>
+                      <Radio
+                        autocomplete="off"
+                        name="suspend-mode"
+                        value={`${!!stream.suspended}`}
+                        checked={!!stream.suspended}
+                      />
+                      <Flex css={{ alignItems: "center" }}>On</Flex>
+                    </Label>
+                    <Label css={{ ml: "0.5em" }}>
+                      <Radio
+                        autocomplete="off"
+                        name="suspend-mode"
+                        value={`${!stream.suspended}`}
+                        checked={!stream.suspended}
+                        onClick={(e) => {
+                          if (stream.suspended) {
+                            setSuspendModal(true);
+                          }
+                        }}
+                      />
+                      <Flex css={{ alignItems: "center" }}>Off</Flex>
+                    </Label> */}
+                        <Flex
+                          css={{
+                            ml: "0.5em",
+                            minWidth: "24px",
+                            height: "24px",
+                            alignItems: "center",
+                          }}>
+                          <Help
+                            data-tip
+                            data-for={`tooltip-suspend-${stream.id}`}
+                            css={{
+                              color: "muted",
+                              cursor: "pointer",
+                              ml: 1,
+                              width: "18px",
+                              height: "18px",
+                            }}
+                          />
+                        </Flex>
+                      </Flex>
+                      <ReactTooltip
+                        id={`tooltip-suspend-${stream.id}`}
+                        className="tooltip"
+                        place="top"
+                        type="dark"
+                        effect="solid">
+                        <p>
+                          When turned on, any active stream sessions will
+                          immediately end.
+                          <br />
+                          New sessions will be prevented from starting until
+                          turned off.
+                        </p>
+                      </ReactTooltip>
+                    </Box>
+                    <Box css={{ m: "0.4em", gridColumn: "1/-1" }}>
+                      <hr />
+                    </Box>
+                    <Cell>Renditions</Cell>
+                    <Cell>
+                      <RenditionsDetails stream={stream} />
+                    </Cell>
+                    <Cell>Created at</Cell>
+                    <Cell>
+                      <RelativeTime
+                        id="cat"
+                        prefix="createdat"
+                        tm={stream.createdAt}
+                        swap={true}
+                      />
+                    </Cell>
+                    <Cell>Last seen</Cell>
+                    <Cell>
+                      <RelativeTime
+                        id="last"
+                        prefix="lastSeen"
+                        tm={stream.lastSeen}
+                        swap={true}
+                      />
+                    </Cell>
+                    <Cell>Status</Cell>
+                    <Cell>{stream.isActive ? "Active" : "Idle"}</Cell>
+                    <Cell>Suspended</Cell>
+                    <Cell>{stream.suspended ? "Suspended" : "Normal"}</Cell>
+                    {user.admin || isStaging() || isDevelopment() ? (
+                      <>
+                        <Cell> </Cell>
+                        <Cell>
+                          <strong>Admin or staging only fields:</strong>
+                        </Cell>
+                      </>
+                    ) : null}
+                    {user.admin ? (
+                      <>
+                        <Cell> </Cell>
+                        <Cell>
+                          <strong>Admin only fields:</strong>
+                        </Cell>
+                        <Cell>Deleted</Cell>
+                        <Cell>
+                          {stream.deleted ? <strong>Yes</strong> : "No"}
+                        </Cell>
+                        <Cell>Source segments</Cell>
+                        <Cell>{stream.sourceSegments || 0}</Cell>
+                        <Cell>Transcoded segments</Cell>
+                        <Cell>{stream.transcodedSegments || 0}</Cell>
+                        <Cell>Source duration</Cell>
+                        <Cell>
+                          {formatNumber(stream.sourceSegmentsDuration || 0, 0)}{" "}
+                          sec (
+                          {formatNumber(
+                            (stream.sourceSegmentsDuration || 0) / 60,
+                            2
+                          )}{" "}
+                          min)
+                        </Cell>
+                        <Cell>Transcoded duration</Cell>
+                        <Cell>
+                          {formatNumber(
+                            stream.transcodedSegmentsDuration || 0,
+                            0
+                          )}{" "}
+                          sec (
+                          {formatNumber(
+                            (stream.transcodedSegmentsDuration || 0) / 60,
+                            2
+                          )}{" "}
+                          min)
+                        </Cell>
+                        <Cell>Source bytes</Cell>
+                        <Cell>{formatNumber(stream.sourceBytes || 0, 0)}</Cell>
+                        <Cell>Transcoded bytes</Cell>
+                        <Cell>
+                          {formatNumber(stream.transcodedBytes || 0, 0)}
+                        </Cell>
+                        <Cell>Ingest rate</Cell>
+                        <Cell>
+                          {formatNumber(stream.ingestRate || 0, 3)} bytes/sec (
+                          {formatNumber((stream.ingestRate || 0) * 8, 0)})
+                          bits/sec
+                        </Cell>
+                        <Cell>Outgoing rate</Cell>
+                        <Cell>
+                          {formatNumber(stream.outgoingRate || 0, 3)} bytes/sec
+                          ({formatNumber((stream.outgoingRate || 0) * 8, 0)})
+                          bits/sec
+                        </Cell>
+                        <Cell>Papertrail to stream key</Cell>
+                        <Cell>
+                          <a
+                            target="_blank"
+                            href={`https://papertrailapp.com/groups/16613582/events?q=${stream.streamKey}`}
+                            css={{ userSelect: "all" }}>
+                            {stream.streamKey}
+                          </a>
+                        </Cell>
+                        <Cell>Papertrail to playback id</Cell>
+                        <Cell>
+                          <a
+                            target="_blank"
+                            href={`https://papertrailapp.com/groups/16613582/events?q=${stream.playbackId}`}
+                            css={{ userSelect: "all" }}>
+                            {stream.playbackId}
+                          </a>
+                        </Cell>
+                        <Cell>Papertrail to stream id</Cell>
+                        <Cell>
+                          <a
+                            target="_blank"
+                            href={`https://papertrailapp.com/groups/16613582/events?q=${stream.id}`}
+                            css={{ userSelect: "all" }}>
+                            {stream.id}
+                          </a>
+                        </Cell>
+                        <Cell>Region/Broadcaster</Cell>
+                        <Cell>
+                          {region}{" "}
+                          {broadcasterHost ? " / " + broadcasterHost : ""}
+                          {stream && stream.mistHost
+                            ? " / " + stream.mistHost
+                            : ""}
+                        </Cell>
+                        {broadcasterPlaybackUrl ? (
+                          <>
+                            <Cell>Broadcaster playback</Cell>
+                            <Cell>
+                              <a
+                                target="_blank"
+                                href={broadcasterPlaybackUrl}
+                                css={{ userSelect: "all" }}>
+                                {broadcasterPlaybackUrl}
+                              </a>
+                            </Cell>
+                          </>
+                        ) : null}
+                      </>
+                    ) : null}
+                  </Box>
+                </Flex>
+                <TimedAlert
+                  text={resultText}
+                  close={() => setResultText("")}
+                  variant="info"
+                />
+                <TimedAlert
+                  text={alertText}
+                  close={() => setAlertText("")}
+                  variant="attention"
+                />
+                <Flex
+                  css={{
+                    justifyContent: "flex-end",
+                    mb: 3,
+                  }}>
+                  {userIsAdmin ? (
+                    <Flex>
+                      <Button
+                        css={{ mr: 3 }}
+                        type="button"
+                        variant="violet"
+                        onClick={() => setTerminateModal(true)}>
+                        Terminate
+                      </Button>
+                    </Flex>
+                  ) : null}
+                  <Button
+                    type="button"
+                    variant="violet"
+                    onClick={() => setDeleteModal(true)}>
+                    Delete
+                  </Button>
+                </Flex>
+                <h1>test</h1>
+                <StreamSessionsTable streamId={stream.id} />
+              </Box>
+              <Box>
+                <Heading size="1" css={{ mb: "$3" }}>
+                  Current Stream
+                </Heading>
+                {videoExists ? (
+                  <Box
+                    css={{
+                      width: "100%",
+                      height: 300,
+                      borderRadius: "$2",
+                      overflow: "hidden",
+                    }}>
+                    <Player
+                      setVideo={setVideoExists}
+                      src={globalPlaybackUrl}
+                      config={{
+                        abr: { enabled: false },
+                        controlPanelElements: [
+                          "time_and_duration",
+                          "play_pause",
+                          "rewind",
+                          "fast_forward",
+                          "mute",
+                          "volume",
+                          "spacer",
+                          "fullscreen",
+                          "overflow_menu",
+                        ],
+                        overflowMenuButtons: ["quality"],
+                      }}
+                    />
+                  </Box>
+                ) : (
+                  <Box
+                    css={{
+                      backgroundColor: "$slate6",
+                      width: "100%",
+                      height: 300,
+                      borderRadius: "$2",
+                    }}
+                  />
+                )}
+
+                <Button
+                  size="3"
+                  variant="violet"
+                  css={{ mt: "$3", width: "100%" }}>
+                  View Stream Health
+                </Button>
+              </Box>
+            </Grid>
+          </>
+        ) : notFound ? (
+          <Box>Not found</Box>
+        ) : (
+          <Flex
+            css={{
+              height: "calc(100vh - 300px)",
+              justifyContent: "center",
+              alignItems: "center",
+            }}>
+            <Spinner />
+          </Flex>
+        )}
+      </Box>
+    </Layout>
+  );
+};
+export default ID;
